refactor(kernel): split mount callback wire
Move bounded callback framing and storage filesystem protocol translation into a coherent sibling module without changing product or test behavior.

Tracking: #1710
